requirements.txt
discord.py[voice]
python-dotenv
requests
yt_dlp